COMMERCE-16 Refactor

+++ modules/dxp/apps/commerce/commerce-product-service/src/main/java/com/liferay/commerce/product/service/impl/CPAttachmentFileEntryLocalServiceImpl.java
@@ -175,7 +175,7 @@ public class CPAttachmentFileEntryLocalServiceImpl
 
 	@Override
 	public List<CPAttachmentFileEntry> getCPAttachmentFileEntries(
-			long classNameId, int classPK, int start, int end)
+			long classNameId, long classPK, int start, int end)
 		throws PortalException {
 
 		return cpAttachmentFileEntryPersistence.findByC_C(
@@ -184,7 +184,7 @@ public class CPAttachmentFileEntryLocalServiceImpl
 
 	@Override
 	public List<CPAttachmentFileEntry> getCPAttachmentFileEntries(
-			long classNameId, int classPK, int start, int end,
+			long classNameId, long classPK, int start, int end,
 			OrderByComparator<CPAttachmentFileEntry> orderByComparator)
 		throws PortalException {
 
@@ -193,7 +193,7 @@ public class CPAttachmentFileEntryLocalServiceImpl
 	}
 
 	@Override
-	public int getCPAttachmentFileEntriesCount(long classNameId, int classPK) {
+	public int getCPAttachmentFileEntriesCount(long classNameId, long classPK) {
 		return cpAttachmentFileEntryPersistence.countByC_C(
 			classNameId, classPK);
 	}
